+++ GRDB Tests/Public/Statement/SelectStatementTests.swift
@@ -62,6 +62,20 @@ class SelectStatementTests : GRDBTestCase {
         }
     }
     
+    func testQueryArgumentsSetterWithArray() {
+        assertNoError {
+            dbQueue.inDatabase { db in
+                let statement = db.selectStatement("SELECT COUNT(*) FROM persons WHERE age < ?")
+                let ages = [20, 30, 40, 50]
+                let counts = ages.map { (age: Int) -> Int in
+                    statement.arguments = [age]
+                    return statement.fetchOne(Int.self)!
+                }
+                XCTAssertEqual(counts, [1,2,2,3])
+            }
+        }
+    }
+    
     func testDictionaryQueryArguments() {
         assertNoError {
             dbQueue.inDatabase { db in
@@ -69,6 +83,21 @@ class SelectStatementTests : GRDBTestCase {
                 // TODO: why is this explicit type declaration required?
                 let ageDicts: [[String: DatabaseValueConvertible?]] = [["age": 20], ["age": 30], ["age": 40], ["age": 50]]
                 let counts = ageDicts.map { statement.fetchOne(Int.self, arguments: QueryArguments($0))! }
+                XCTAssertEqual(counts, [1,2,2,3])
+            }
+        }
+    }
+    
+    func testQueryArgumentsSetterWithDictionary() {
+        assertNoError {
+            dbQueue.inDatabase { db in
+                let statement = db.selectStatement("SELECT COUNT(*) FROM persons WHERE age < :age")
+                // TODO: why is this explicit type declaration required?
+                let ageDicts: [[String: DatabaseValueConvertible?]] = [["age": 20], ["age": 30], ["age": 40], ["age": 50]]
+                let counts = ageDicts.map { (ageDict: [String: DatabaseValueConvertible?]) -> Int in
+                    statement.arguments = QueryArguments(ageDict)
+                    return statement.fetchOne(Int.self)!
+                }
                 XCTAssertEqual(counts, [1,2,2,3])
             }
         }